+++ component-resources.js
@@ -7,7 +7,19 @@
   const version = loaderUrl.searchParams.get('ver');
   const resources = window.__resources || {};
 
-  for (const file of ['StorySection.dc.html', 'TicketsSection.dc.html']) {
+  const componentFiles = [
+    'HeaderSection.dc.html',
+    'HeroSection.dc.html',
+    'ReleaseSection.dc.html',
+    'StorySection.dc.html',
+    'TicketsSection.dc.html',
+    'VideosSection.dc.html',
+    'ShortsSection.dc.html',
+    'CastSection.dc.html',
+    'FooterSection.dc.html'
+  ];
+
+  for (const file of componentFiles) {
     const componentUrl = new URL(file, loaderUrl);
     if (version) componentUrl.searchParams.set('ver', version);
     resources[`./${file}`] = componentUrl.href;